Refactor database checks at installation
